Refactored the Netty stuff to a separate util class
